fix: ajusta tipo coluna code

+++ database/migrations/2026_01_09_181910_create_states_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('states', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->integer('code');
+            $table->bigInteger('code');
             $table->string('acronym');
             $table->timestamps();
         });
